+++ apps/web/src/app/[locale]/quiz/_components/BackGround.tsx
