+++ apps/web/src/app/[locale]/quiz/_components/CreateOrSolve/QuizTypeCard.tsx
