+++ apps/web/src/app/[locale]/quiz/_components/CreateOrSolve/SolveQuizConfirmDialog.tsx
