+++ apps/web/src/app/[locale]/quiz/create/_components/QuizTextArea.tsx
