+++ apps/web/src/app/[locale]/quiz/layout.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
 
-import { MobileLayout } from '@/app/[locale]/quiz/_common/MobileLayout';
+import { MobileLayout } from '@/app/[locale]/quiz/_components/MobileLayout';
 import GNB from '@/components/GNB/GNB';
 
 export default function QuizLayout({ children }: { children: React.ReactNode }) {
