+++ apps/web/src/app/[locale]/quiz/solve/_components/done/CompleteAlertDialog.tsx
