+++ apps/web/src/app/[locale]/quiz/solve/_components/fail/FailAlertDialog.tsx
