+++ apps/web/src/app/[locale]/quiz/solve/_components/success/CorrectConfirmDialog.tsx
