+++ apps/web/src/app/[locale]/quiz/solve/_hooks/useQuizAction.ts
@@ -3,17 +3,22 @@
 import { useRouter } from 'next/navigation';
 import { toast } from 'sonner';
 
-import type useQuizDialogStatus from '@/app/[locale]/quiz/solve/(status)/_solving/useQuizDialogStatus';
-import type { QuizAnswer } from '@/app/[locale]/quiz/solve/solveQuiz.constants';
-import { QUIZ_ANSWER, QUIZ_POINT_MAP, QUIZ_TOTAL_STAGE } from '@/app/[locale]/quiz/solve/solveQuiz.constants';
-import { useSolveQuizContext } from '@/app/[locale]/quiz/solve/SolveQuizContext';
+import type { QuizAnswer } from '@/app/[locale]/quiz/solve/_constants/solveQuiz.constants';
+import {
+  QUIZ_ANSWER,
+  QUIZ_POINT_MAP,
+  QUIZ_TOTAL_STAGE,
+} from '@/app/[locale]/quiz/solve/_constants/solveQuiz.constants';
+
+import type useQuizDialogStatus from './useQuizDialogStatus';
 
 interface UseQuizActionProps {
+  stage: number;
+  setStage: (stage: number) => void;
   quizDialog: ReturnType<typeof useQuizDialogStatus>;
 }
 
-const useQuizAction = ({ quizDialog }: UseQuizActionProps) => {
-  const { stage, setStage } = useSolveQuizContext();
+const useQuizAction = ({ stage, setStage, quizDialog }: UseQuizActionProps) => {
   const { correctDialog, failDialog, completeDialog } = quizDialog;
 
   const submit = (answer: QuizAnswer) => {
